testheroku/requirements/common.txt
Django==2.2.4
django-decouple==2.1
Pillow==6.1.0
pytz==2019.2
sqlparse==0.3.0
psycopg2-binary==2.8.3
dj-static==0.0.6
